+++ source/Application/Controller/Admin/GenericExportDo.php
@@ -6,7 +6,8 @@
 
 namespace OxidEsales\EshopCommunity\Application\Controller\Admin;
 
-use oxRegistry;
+use OxidEsales\EshopCommunity\Internal\Templating\TemplateRendererBridgeInterface;
+use OxidEsales\EshopCommunity\Internal\Templating\TemplateRendererInterface;
 
 /**
  * General export class.
@@ -54,18 +55,38 @@ class GenericExportDo extends \OxidEsales\Eshop\Application\Controller\Admin\Dyn
         $blContinue = false;
         if ($oArticle = $this->getOneArticle($iCnt, $blContinue)) {
             $myConfig = \OxidEsales\Eshop\Core\Registry::getConfig();
-            $oSmarty = \OxidEsales\Eshop\Core\Registry::getUtilsView()->getSmarty();
-            $oSmarty->assign("sCustomHeader", \OxidEsales\Eshop\Core\Registry::getSession()->getVariable("sExportCustomHeader"));
-            $oSmarty->assign("linenr", $iCnt);
-            $oSmarty->assign("article", $oArticle);
-            $oSmarty->assign("spr", $myConfig->getConfigParam('sCSVSign'));
-            $oSmarty->assign("encl", $myConfig->getConfigParam('sGiCsvFieldEncloser'));
-            $this->write($oSmarty->fetch("genexport.tpl", $this->getViewId()));
+            $context = [
+                "sCustomHeader" => \OxidEsales\Eshop\Core\Registry::getSession()->getVariable("sExportCustomHeader"),
+                "linenr"        => $iCnt,
+                "article"       => $oArticle,
+                "spr"           => $myConfig->getConfigParam('sCSVSign'),
+                "encl"          => $myConfig->getConfigParam('sGiCsvFieldEncloser')
+            ];
+            $context['oxEngineTemplateId'] = $this->getViewId();
+
+            $this->write(
+                $this->getRenderer()->renderTemplate(
+                    "genexport.tpl",
+                    $context
+                )
+            );
 
             return ++$iExportedItems;
         }
 
         return $blContinue;
+    }
+
+    /**
+     * @internal
+     *
+     * @return TemplateRendererInterface
+     */
+    private function getRenderer()
+    {
+        return $this->getContainer()
+            ->get(TemplateRendererBridgeInterface::class)
+            ->getTemplateRenderer();
     }
 
     /**