+++ source/Core/WidgetControl.php
@@ -8,6 +8,8 @@ namespace OxidEsales\EshopCommunity\Core;
 
 use OxidEsales\Eshop\Application\Component\Widget\WidgetController;
 use OxidEsales\Eshop\Core\Exception\ObjectException;
+use OxidEsales\EshopCommunity\Internal\Templating\TemplateRendererBridgeInterface;
+use OxidEsales\EshopCommunity\Internal\Templating\TemplateRendererInterface;
 
 /**
  * Main shop actions controller. Processes user actions, logs
@@ -81,8 +83,8 @@ class WidgetControl extends \OxidEsales\Eshop\Core\ShopControl
             }
 
             // Setting back last active view.
-            $oSmarty = \OxidEsales\Eshop\Core\Registry::getUtilsView()->getSmarty();
-            $oSmarty->assign('oView', $oConfig->getActiveView());
+            $engine = $this->getRenderer()->getTemplateEngine();
+            $engine->addGlobal('oView', $oConfig->getActiveView());
         }
     }
 
@@ -137,5 +139,17 @@ class WidgetControl extends \OxidEsales\Eshop\Core\ShopControl
         }
 
         return $widgetViewObject;
+    }
+
+    /**
+     * @internal
+     *
+     * @return TemplateRendererInterface
+     */
+    private function getRenderer()
+    {
+        return $this->getContainer()
+            ->get(TemplateRendererBridgeInterface::class)
+            ->getTemplateRenderer();
     }
 }