+++ source/Internal/Smarty/Configuration/SmartySecuritySettingsDataProviderInterface.php
@@ -1,0 +1,21 @@
+<?php declare(strict_types=1);
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace OxidEsales\EshopCommunity\Internal\Smarty\Configuration;
+
+/**
+ * Interface SmartySecuritySettingsDataProviderInterface
+ * @package OxidEsales\EshopCommunity\Internal\Smarty\Configuration
+ */
+interface SmartySecuritySettingsDataProviderInterface
+{
+    /**
+     * Return smarty security settings.
+     *
+     * @return array
+     */
+    public function getSecuritySettings(): array;
+}