+++ source/Internal/Templating/TemplateEngineFactoryInterface.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace OxidEsales\EshopCommunity\Internal\Templating;
+
+/**
+ * Interface TemplateEngineFactoryInterface
+ * @package OxidEsales\EshopCommunity\Internal\Templating
+ */
+interface TemplateEngineFactoryInterface
+{
+    /**
+     * @return TemplateEngineInterface
+     */
+    public function getTemplateEngine(): TemplateEngineInterface;
+}