+++ tests/Unit/Application/Controller/ShopControlTest.php
@@ -13,11 +13,14 @@ use OxidEsales\Eshop\Core\Registry;
 use OxidEsales\EshopCommunity\Core\Exception\ConnectionException;
 use OxidEsales\EshopCommunity\Core\Exception\ExceptionToDisplay;
 use OxidEsales\EshopCommunity\Core\Output;
+use OxidEsales\EshopCommunity\Internal\Application\ContainerFactory;
+use OxidEsales\EshopCommunity\Internal\Templating\TemplateRendererInterface;
 use oxOutput;
 use oxRegistry;
 use oxSystemComponentException;
 use oxTestModules;
 use Psr\Log\LoggerInterface;
+use Psr\Container\ContainerInterface;
 
 // Force autoloading of Smarty class, so that mocking would work correctly.
 class_exists('Smarty');
@@ -262,6 +265,7 @@ class ShopControlTest extends \OxidTestCase
      */
     public function testRenderTemplateNotFound()
     {
+        ContainerFactory::resetContainer();
         $oView = $this->getMock(\OxidEsales\Eshop\Core\Controller\BaseController::class, array('render'));
         $oView->expects($this->once())->method('render')->will($this->returnValue('wrongTpl'));
 
@@ -275,7 +279,9 @@ class ShopControlTest extends \OxidTestCase
         $oControl->expects($this->any())->method('_isDebugMode')->will($this->returnValue(true));
 
         $oSmarty = $this->getMock("Smarty", array('fetch'));
-        $oSmarty->expects($this->once())->method('fetch')->with($this->equalTo("message/exception.tpl"));
+        $oSmarty->expects($this->once())->method('fetch')
+            ->with($this->equalTo("message/exception.tpl"))
+            ->will($this->returnValue(''));
 
         $oUtilsView = $this->getMock(\OxidEsales\Eshop\Core\UtilsView::class, array('getSmarty'));
         $oUtilsView->expects($this->once())->method('getSmarty')->will($this->returnValue($oSmarty));
@@ -300,6 +306,7 @@ class ShopControlTest extends \OxidTestCase
         if ($this->getTestConfig()->getShopEdition() == 'EE') {
             $this->markTestSkipped('This test is for Community/Professional edition only.');
         }
+        ContainerFactory::resetContainer();
         $this->getConfig()->setConfigParam('sTheme', 'azure');
 
         $controllerClassName = 'content';
@@ -333,6 +340,7 @@ class ShopControlTest extends \OxidTestCase
         if ($this->getTestConfig()->getShopEdition() == 'EE') {
             $this->markTestSkipped('This test is for Community/Professional edition only.');
         }
+        ContainerFactory::resetContainer();
         $this->getConfig()->setConfigParam('sTheme', 'azure');
 
         $controllerClassName = 'content';
@@ -370,6 +378,7 @@ class ShopControlTest extends \OxidTestCase
         if ($this->getTestConfig()->getShopEdition() == 'EE') {
             $this->markTestSkipped('This test is for Community/Professional edition only.');
         }
+        ContainerFactory::resetContainer();
         $this->getConfig()->setConfigParam('sTheme', 'azure');
 
         $controllerClassName = 'content';
@@ -618,7 +627,9 @@ class ShopControlTest extends \OxidTestCase
     private function getSmartyMock($expectedTemplate)
     {
         $oSmarty = $this->getMock("Smarty", array('fetch'));
-        $oSmarty->expects($this->once())->method('fetch')->with($this->equalTo($expectedTemplate));
+        $oSmarty->expects($this->once())->method('fetch')
+            ->with($this->equalTo($expectedTemplate))
+            ->will($this->returnValue('string'));
 
         return $oSmarty;
     }